REST API: Prevent database error when deleting meta data.
Add a check to `WP_REST_Meta_Fields::delete_meta_value()` ensuring meta data is set before attempting to delete it from the database. If the data does not exist, the delete is considered successful as the data matches the desired state.

Props BrechtVds, goaroundagain, TimothyBlynJacobs.
Merges [50567] to the 5.7 branch.
Fixes #52787.

Built from https://develop.svn.wordpress.org/branches/5.7@50573


git-svn-id: http://core.svn.wordpress.org/branches/5.7@50186 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-includes/rest-api/fields/class-wp-rest-meta-fields.php
@@ -232,6 +232,10 @@ abstract class WP_REST_Meta_Fields {
 			);
 		}
 
+		if ( null === get_metadata_raw( $meta_type, $object_id, wp_slash( $meta_key ) ) ) {
+			return true;
+		}
+
 		if ( ! delete_metadata( $meta_type, $object_id, wp_slash( $meta_key ) ) ) {
 			return new WP_Error(
 				'rest_meta_database_error',
